Ignore hidden files when cleaning output dir
A common pattern for deploying to GitHub pages is
to set up nested git checkouts as follows:

project/ - Checkout of 'source' branch of project
project/_site - Checkout of 'master' branch of project
                with generated content

Running the generator in 'project/' updates the 'project/_site'
folder which is then easily checked in as the deployed branch.

+++ fs_util.ts
@@ -2,11 +2,13 @@ import fs = require('fs');
 import fs_extra = require('fs-extra');
 import path = require('path');
 
-export function cleanDir(dir: string) {
+export function cleanDir(dir: string, filter?: (path: string) => boolean) {
 	var files = fs.readdirSync(dir);
 	files.forEach((file) => {
 		var filePath = path.resolve(`${dir}/${file}`);
-		fs_extra.removeSync(filePath);
+		if (!filter || filter(filePath)) {
+			fs_extra.removeSync(filePath);
+		}
 	});
 }
